VetGuard/requirements.txt
fastapi==0.115.12
gdown==5.2.0
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.18.0
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==0.21.0
google-genai==1.23.0
